Compiler warnings due to gcc differences (from Diego Quagreda)
git-svn-id: https://svn.code.sf.net/p/tcnopen/trdp/trunk@613 3b5a3598-5f4e-4449-9e63-bd40438bfec0

+++ trdp/src/vos/api/vos_utils.h
@@ -82,7 +82,7 @@ extern void *gRefCon;
     #define ALIGNOF(type)   __alignof__(type)
 */
 #else
-    #define ALIGNOF(type)  offsetof(struct { char c; type member; }, member)
+    #define ALIGNOF(type)  ((UINT32)offsetof(struct { char c; type member; }, member))
 #endif
 
 /***********************************************************************************************************************
